refactor: Phase 3.1f — extract GlassIntegrationMixin from ElectrodeAdvisorWidget
Move 9 glass integration methods (209 lines) into
GlassIntegrationMixin in main_window_glass_integration.py:
- set_glass_properties_calculator
- connect_to_glass_calculator
- _on_glass_integration_changed
- _check_glass_calculator_availability
- _find_main_window
- _connect_to_glass_calculator
- _update_glass_integration_status
- _on_glass_properties_updated

Monolith now at 1,535 lines (-65.0% from original 4,387).

7 mixins extracted total.
